Update trakt.py
In general: Added functionality to limit the results gathered from Trakt lists. (ie. You can just grab the top x from the list, rather than the whole list.)

Expanded non_user_lists to include 'recommended'. I realize this is a user list of sorts, but it is handled throughout the script like a non_user_list.

Added cache_time class attribute to Class Trakt so that it can be changed from within Traktarr.py for limited results searches.

Added a function to return the factor that is both closest to and under 1000, and it's opposing factor. Utilizing this to determine how to divide the results_limit between pages*results_per_page.

Trakt's Box Office Movies list cannot be limited for whatever reason. There is some handling in there in case the user tries to limit it from CLI or the automatic configuration.

Added the passing of the 'page' argument to various functions.